+++ MatCalcitePlugin/src/com/github/vlsi/mat/calcite/functions/TableFunctions.java
@@ -1,9 +1,11 @@
 package com.github.vlsi.mat.calcite.functions;
 
+import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 
 import org.apache.calcite.adapter.java.AbstractQueryableTable;
 import org.apache.calcite.linq4j.BaseQueryable;
@@ -23,8 +25,10 @@ import org.apache.calcite.sql.type.SqlTypeName;
 import org.apache.calcite.util.ImmutableBitSet;
 import org.eclipse.mat.SnapshotException;
 import org.eclipse.mat.inspections.collectionextract.CollectionExtractionUtils;
+import org.eclipse.mat.inspections.collectionextract.ExtractedMap;
 import org.eclipse.mat.inspections.collectionextract.ICollectionExtractor;
 import org.eclipse.mat.snapshot.ISnapshot;
+import org.eclipse.mat.snapshot.model.IObject;
 import org.eclipse.mat.snapshot.model.NamedReference;
 import org.eclipse.mat.util.VoidProgressListener;
 
@@ -39,6 +43,7 @@ public class TableFunctions {
     public static Multimap<String, TableFunction> createAll() {
         ImmutableMultimap.Builder<String, TableFunction> builder = ImmutableMultimap.builder();
         builder.put("getValues", TableFunctionImpl.create(TableFunctions.class, "getValues"));
+        builder.put("getMapEntries", TableFunctionImpl.create(TableFunctions.class, "getMapEntries"));
         builder.put("getRetainedSet", TableFunctionImpl.create(TableFunctions.class, "getRetainedSet"));
         builder.put("getOutboundReferences", TableFunctionImpl.create(TableFunctions.class, "getOutboundReferences"));
         builder.put("getInboundReferences", TableFunctionImpl.create(TableFunctions.class, "getInboundReferences"));
@@ -56,8 +61,7 @@ public class TableFunctions {
                 ICollectionExtractor collectionExtractor = CollectionExtractionUtils.findCollectionExtractor(ref.getIObject());
                 if (collectionExtractor == null) {
                     references = Collections.emptyList();
-                }
-                else {
+                } else {
                     references = collectReferences(ref.getIObject().getSnapshot(), collectionExtractor.extractEntryIds(ref.getIObject()));
                 }
             } catch (SnapshotException e) {
@@ -68,12 +72,39 @@ public class TableFunctions {
     }
 
     @SuppressWarnings("unused")
+    public static QueryableTable getMapEntries(Object r) {
+        List<HeapReference[]> references;
+        if (!(r instanceof HeapReference)) {
+            references = Collections.emptyList();
+        } else {
+            HeapReference ref = (HeapReference) r;
+            try {
+                ExtractedMap extractedMap = CollectionExtractionUtils.extractMap(ref.getIObject());
+                if (extractedMap == null) {
+                    references = Collections.emptyList();
+                } else {
+                    references = new ArrayList<>();
+                    for (Map.Entry<IObject, IObject> entry : extractedMap) {
+                        references.add(new HeapReference[]{
+                                HeapReference.valueOf(entry.getKey()),
+                                HeapReference.valueOf(entry.getValue())
+                        });
+                    }
+                }
+            } catch (SnapshotException e) {
+                throw new RuntimeException("Cannot extract values from " + ref, e);
+            }
+        }
+        return new HeapReferencesTable(new String[]{"key", "value"}, references);
+    }
+
+    @SuppressWarnings("unused")
     public static QueryableTable getRetainedSet(Object r) {
         List<HeapReference> references;
         if (!(r instanceof HeapReference)) {
             references = Collections.emptyList();
         } else {
-            HeapReference ref = (HeapReference)r;
+            HeapReference ref = (HeapReference) r;
             ISnapshot snapshot = ref.getIObject().getSnapshot();
             try {
                 references = collectReferences
@@ -82,7 +113,7 @@ public class TableFunctions {
                                 snapshot.getRetainedSet(new int[]{ref.getIObject().getObjectId()}, new VoidProgressListener())
                         );
             } catch (SnapshotException e) {
-                throw new RuntimeException("Cannot extract retained set from "+r, e);
+                throw new RuntimeException("Cannot extract retained set from " + r, e);
             }
         }
         return new HeapReferenceTable(references, true);
@@ -94,7 +125,7 @@ public class TableFunctions {
         if (!(r instanceof HeapReference)) {
             references = Collections.emptyList();
         } else {
-            HeapReference ref = (HeapReference)r;
+            HeapReference ref = (HeapReference) r;
             references = ref.getIObject().getOutboundReferences();
         }
         return new OutboundReferencesTable(references);
@@ -106,7 +137,7 @@ public class TableFunctions {
         if (!(r instanceof HeapReference)) {
             references = Collections.emptyList();
         } else {
-            HeapReference ref = (HeapReference)r;
+            HeapReference ref = (HeapReference) r;
             ISnapshot snapshot = ref.getIObject().getSnapshot();
             try {
                 references = collectReferences
@@ -115,7 +146,7 @@ public class TableFunctions {
                                 snapshot.getInboundRefererIds(ref.getIObject().getObjectId())
                         );
             } catch (SnapshotException e) {
-                throw new RuntimeException("Cannot extract inbound references from "+r, e);
+                throw new RuntimeException("Cannot extract inbound references from " + r, e);
             }
         }
         return new HeapReferenceTable(references, true);
@@ -124,7 +155,7 @@ public class TableFunctions {
     private static List<HeapReference> collectReferences(ISnapshot snapshot, int[] objectIds) throws SnapshotException {
         if (objectIds != null && objectIds.length > 0) {
             List<HeapReference> references = new ArrayList<>();
-            for(int objectId : objectIds) {
+            for (int objectId : objectIds) {
                 references.add(HeapReference.valueOf(snapshot.getObject(objectId)));
             }
             return references;
@@ -133,26 +164,40 @@ public class TableFunctions {
         }
     }
 
-    private static class HeapReferenceTable extends AbstractQueryableTable {
+    private static class HeapReferenceTable extends ValuesListTable<HeapReference> {
+        HeapReferenceTable(Collection<HeapReference> references, boolean unique) {
+            super(HeapReference.class, new String[]{"this"}, references, unique);
+        }
+    }
+
+    private static class HeapReferencesTable extends ValuesListTable<HeapReference[]> {
+        HeapReferencesTable(String[] columnNames, Collection<HeapReference[]> references) {
+            super(HeapReference[].class, columnNames, references, false);
+        }
+    }
+
+    private static class ValuesListTable<RowType> extends AbstractQueryableTable {
         private final static List<ImmutableBitSet> UNIQUE_KEYS_STATISTICS = ImmutableList.of(ImmutableBitSet.of(0));
         private final static List<ImmutableBitSet> NON_UNIQUE_KEYS_STATISTICS = ImmutableList.of(ImmutableBitSet.of());
 
-        private final Collection<HeapReference> references;
+        private final String[] columnNames;
+        private final Collection<RowType> values;
         private final boolean unique;
 
-        public HeapReferenceTable(Collection<HeapReference> references, boolean unique) {
-            super(HeapReference.class);
-            this.references = references;
+        ValuesListTable(Type targetRowType, String[] columnNames, Collection<RowType> values, boolean unique) {
+            super(targetRowType);
+            this.columnNames = columnNames;
+            this.values = values;
             this.unique = unique;
         }
 
         @SuppressWarnings("unchecked")
         @Override
         public <T> Queryable<T> asQueryable(QueryProvider queryProvider, SchemaPlus schemaPlus, String s) {
-            BaseQueryable<HeapReference> queryable = new BaseQueryable<HeapReference>(null, HeapReference.class, null) {
+            BaseQueryable<RowType> queryable = new BaseQueryable<RowType>(null, getElementType(), null) {
                 @Override
-                public Enumerator<HeapReference> enumerator() {
-                    return Linq4j.enumerator(references);
+                public Enumerator<RowType> enumerator() {
+                    return Linq4j.enumerator(values);
                 }
             };
             return (Queryable<T>) queryable;
@@ -160,12 +205,16 @@ public class TableFunctions {
 
         @Override
         public RelDataType getRowType(RelDataTypeFactory relDataTypeFactory) {
-            return relDataTypeFactory.builder().add("this",  relDataTypeFactory.createSqlType(SqlTypeName.ANY)).build();
+            RelDataTypeFactory.FieldInfoBuilder builder = relDataTypeFactory.builder();
+            for (String columnName : columnNames) {
+                builder.add(columnName, relDataTypeFactory.createSqlType(SqlTypeName.ANY));
+            }
+            return builder.build();
         }
 
         @Override
         public Statistic getStatistic() {
-            return Statistics.of(references.size(), unique ? UNIQUE_KEYS_STATISTICS : NON_UNIQUE_KEYS_STATISTICS);
+            return Statistics.of(values.size(), unique ? UNIQUE_KEYS_STATISTICS : NON_UNIQUE_KEYS_STATISTICS);
         }
     }
 }