+++ MatCalcitePlugin/src/com/github/vlsi/mat/calcite/neo/PackageSchema.java
@@ -1,5 +1,6 @@
 package com.github.vlsi.mat.calcite.neo;
 
+import com.github.vlsi.mat.calcite.functions.MapFunctions;
 import com.github.vlsi.mat.calcite.schema.objects.IClassesList;
 import com.github.vlsi.mat.calcite.schema.objects.InstanceByClassTable;
 import com.github.vlsi.mat.calcite.schema.objects.InstanceIdsByClassTable;
@@ -96,6 +97,7 @@ public class PackageSchema extends AbstractSchema {
             // Create functions for schema
             ImmutableMultimap.Builder<String, Function> builder = ImmutableMultimap.builder();
             builder.putAll(ScalarFunctionImpl.createAll(HeapFunctions.class));
+            builder.putAll(MapFunctions.createAll());
             builder.putAll(TableFunctions.createAll());
             builder.putAll(SnapshotFunctions.createAll(snapshot));
             builder.put("getHints", new SqlAdvisorGetHintsFunction());